# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is helper API for IoT and other things.",
        "title": "Homestack API",
        "contact": {},
        "version": "1.0"
    },
    "host": "192.168.0.123:5001",
    "basePath": "/api/v1",
    "paths": {
        "/docker-info": {
            "get": {
                "description": "shows docker containers/images basic information.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker"
                ],
                "summary": "Show docker images or containers.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "item type: containers or images",
                        "name": "items",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/docker-logs": {
            "get": {
                "description": "shows docker logs of running container.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "docker"
                ],
                "summary": "Show docker container logs.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "container name",
                        "name": "container",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "since when in minutes",
                        "name": "since",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/redis-info": {
            "get": {
                "description": "shows vibration-sensor, door-state. humiditym and washing-state.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redis"
                ],
                "summary": "Show most used keys from Redis.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}